+++ tests/testthat/test_functions-MSpectra.R
@@ -81,3 +81,29 @@ test_that("show,MSpectra works", {
     .show_MSpectra(spl, print.classinfo = TRUE)
     show(spl)
 })
+
+test_that("extractSpectraData works", {
+    fl <- system.file("TripleTOF-SWATH", "PestMix1_SWATH.mzML",
+                      package = "msdata")
+    data <- filterRt(readMSData(fl, mode = "onDisk"), rt = c(1, 6))
+    sps <- spectra(data)
+
+    res <- extractSpectraData(sps)
+    expect_true(is(res, "DataFrame"))
+    expect_true(all(c("mz", "intensity") %in% colnames(res)))
+    expect_true(is(res$mz, "NumericList"))
+    expect_true(is(res$intensity, "NumericList"))
+
+    expect_error(extractSpectraData(1:10), "should be either a 'list'")
+
+    res <- extractSpectraData(data)
+    expect_true(is(res, "DataFrame"))
+    expect_true(all(c("mz", "intensity") %in% colnames(res)))
+    expect_true(is(res$mz, "NumericList"))
+    expect_true(is(res$intensity, "NumericList"))
+
+    spctra <- MSpectra(sps)
+    mcols(spctra)$new_col <- "a"
+    res <- extractSpectraData(sps)
+    expect_true(all(res$new_col == "a"))
+})